% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFigures.R
\name{compareFigures}
\alias{compareFigures}
\title{Generate diff between figures}
\usage{
compareFigures(.path)
}
\arguments{
\item{.path}{file path to a figure (pdf or png) or a directory of figures}
}
\description{
Allows the user to compare the differences between figures. The user can
input either a directory of figures or a singular figure and a HTML page
will be generated showing the last version of the figure checked into SVN
next to the local version.
}
