% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModified.R
\name{compareModified}
\alias{compareModified}
\title{Generate comparison between local (modified) and checked in file(s)}
\usage{
compareModified(.path)
}
\arguments{
\item{.path}{file or directory path to tables/figures of interest (pdf, png, or tex)}
}
\description{
Generate comparison between local (modified) and checked in file(s)
}
