% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFigures.R
\name{compareModifiedFigures}
\alias{compareModifiedFigures}
\title{Generate comparison between local (modified) and checked in figures}
\usage{
compareModifiedFigures(.path)
}
\arguments{
\item{.path}{file path to a directory of figures or a singular figure (pdf or png)}
}
\description{
Allows the user to compare their local version of a figure to the checked
in version in SVN. The user can input either a directory of figures or a
singular figure and a HTML page will be generated showing the last version
of the figure checked into SVN next to the local version.
}
