% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTables.R
\name{compareTables}
\alias{compareTables}
\title{Generate comparison between sets of tables}
\usage{
compareTables(.path_current, .path_previous = NULL, .side_by_side = TRUE)
}
\arguments{
\item{.path_current}{file or directory path to tables of interest}

\item{.path_previous}{.path_previous file or directory path to compare .path_base to (leave as NULL to compare to repo version)}

\item{.side_by_side}{Logical. Should outputs be displayed side by side?}
}
\description{
Generate a html document showing two versions of a figure (tex).
If two paths are provided, the local versions of both will be used.
If only one path is provided, the comparison will show the local
versus last checked in version.
}
