% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffFigures.R
\name{diffFigures}
\alias{diffFigures}
\title{Generate diff between figures}
\usage{
diffFigures(figure1, figure2 = NULL)
}
\arguments{
\item{figure1}{file path to figure 1}

\item{figure2}{file path to figure 2 (optional)}
}
\description{
Allows the user to compare the differences between two figures or different
versions of the same figure. If file paths are provided to both the \code{figure1}
and \code{figure2} argument, a HTML page will be generated showing both side by
side.

If only the \code{figure1} argument is provided, a HTML page will be generated
showing the last version of the figure checked into SVN next to the
local version.
}
