% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffFiles.R
\name{diffFiles}
\alias{diffFiles}
\title{Visual diff between two different files}
\usage{
diffFiles(
  .file_current,
  .file_previous,
  .current_revision_header = "Current file",
  .previous_revision_header = "Previous file"
)
}
\arguments{
\item{.file_current}{file path of new file from working directory}

\item{.file_previous}{file path of previous file from working directory}

\item{.current_revision_header}{Header for current file in viewer (default: "Current file")}

\item{.previous_revision_header}{Header for current file in viewer (default: "Previous file")}
}
\description{
Compares the current version of two scripts against one another.
The output will appear in the viewer and only rows where there have been
additions, deletions or modifications in the script will be shown.

This should be used when a new script has been created based off
an existing script.
}
\examples{
\dontrun{
diffFiles(.file_current = "script/data-assembly-v2.Rmd", 
          .file_previous = "script/data-assembly-v1.Rmd")
}

}
