% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffNewFile.R
\name{diffNewFile}
\alias{diffNewFile}
\title{Visual diff between two different files}
\usage{
diffNewFile(.file_new, .file_previous)
}
\arguments{
\item{.file_new}{file path of new file from working directory}

\item{.file_previous}{file path of previous file from working directory}
}
\description{
Compares the current version of two scripts against one another.
The output will appear in the viewer and only rows where there have been
additions, deletions or modifications in the script will be shown.

This should be used when a new script has been created based off
an existing script.
}
\examples{
\dontrun{
diffNewFile(.file_new = "script/data-assembly-v2.Rmd", .file_previous = "script/data-assembly-v1.Rmd")
}

}
