% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffPDF.R
\name{diffPDF}
\alias{diffPDF}
\title{Create a GIF of original vs modified PDF}
\usage{
diffPDF(.file, .fps = 1)
}
\arguments{
\item{.file}{file path from working directory}

\item{.fps}{frames per second, increase for quicker switching between images}
}
\description{
Compares the local version of a script with the most recent QCed version.
The output will appear in the viewer and only rows where there have been
additions, deletions or modifications in the script will be shown.
}
\examples{
\dontrun{
diffPDF(.file = "example-1.pdf", .fps = 1)
}

}
