% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffPreviousRevisions.R
\name{diffPreviousRevisions}
\alias{diffPreviousRevisions}
\title{Visual diff of previous revision of a file to the local version.}
\usage{
diffPreviousRevisions(
  .file,
  .previous_revision,
  .current_revision = NULL,
  .side_by_side = TRUE,
  .ignore_white_space = FALSE
)
}
\arguments{
\item{.file}{file path from working directory}

\item{.previous_revision}{revision to compare to current revision}

\item{.current_revision}{current revision (defaults to local copy)}

\item{.side_by_side}{Logical. Should diffs be displayed side by side?}

\item{.ignore_white_space}{Logical. Should white space be ignored?}
}
\description{
Compares the current version of a script with a previous revision.
The output will appear in the viewer and only rows where there have been
additions, deletions or modifications in the script will be shown.
}
\examples{
with_demoRepo({
 diffPreviousRevisions(.file = "script/data-assembly.R", 
                       .previous_revision = 2)
})

}
