% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffQced.R
\name{diffQced}
\alias{diffQced}
\title{Visual diff of last QCed version of a file to the local version.}
\usage{
diffQced(
  .file,
  .side_by_side = TRUE,
  .ignore_white_space = FALSE,
  .display_entire_file = FALSE,
  .open_file = FALSE
)
}
\arguments{
\item{.file}{file path from working directory}

\item{.side_by_side}{Logical. Should diffs be displayed side by side?}

\item{.ignore_white_space}{Logical. Should white space be ignored?}

\item{.display_entire_file}{Logical. Should the entire file be displayed?}

\item{.open_file}{Logical. Should the file open after generating diff?}
}
\description{
Compares the local version of a script with the most recent QCed version.
The output will appear in the viewer and only rows where there have been
additions, deletions or modifications in the script will be shown.
}
\details{
The file must be part of an SVN working copy that is in sync with the repo.
If the local copy is out of date or has local modifications, the function
will stop and prompt you to update or review local changes.
}
\examples{
with_demoRepo({
  diffQced("script/data-assembly.R")
})

}
