% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffSinceMyLastEdit.R
\name{diffSinceMyLastEdit}
\alias{diffSinceMyLastEdit}
\title{Visual diff of changes to a file since user's last modifications.}
\usage{
diffSinceMyLastEdit(
  .file,
  .host_name = "mc1.metrumrg.com",
  .side_by_side = TRUE,
  .ignore_white_space = FALSE
)
}
\arguments{
\item{.file}{file path from working directory}

\item{.host_name}{Host name where repository is hosted}

\item{.side_by_side}{Logical. Should diffs be displayed side by side?}

\item{.ignore_white_space}{Logical. Should white space be ignored?}
}
\description{
Compares the local version of a script with the most recent QCed version.
The output will appear in the viewer and only rows where there have been
additions, deletions or modifications in the script will be shown.
}
\examples{
\dontrun{
diffSinceMyLastEdit(.file = "script/data-assembly.Rmd")
}

}
