% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileDelete.R
\name{fileDelete}
\alias{fileDelete}
\title{Delete file in SVN and QClog}
\usage{
fileDelete(.filepath)
}
\arguments{
\item{.filepath}{current file path of file to be deleted}
}
\description{
This function facilitates deleting a file in SVN. When run
the file will be deleted and removed from the QC log. A
commit will be made with a message describing the updates.
}
