% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileRename.R
\name{fileRename}
\alias{fileRename}
\title{Rename file in SVN and QClog}
\usage{
fileRename(.filepath, .new_filepath)
}
\arguments{
\item{.filepath}{current file path of file to be renamed}

\item{.new_filepath}{new file path for file}
}
\description{
This function facilitates renaming a file in SVN. When run
the file will be renamed and a targeted commit to SVN will
be made to check in the new name. Additionally, if the file
exists in the QC log, all entries will be renamed and the updated
QC log will also be checked in to SVN.
}
