% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMyLastEditRev.R
\name{getMyLastEditRev}
\alias{getMyLastEditRev}
\title{Find revision of user's last modification}
\usage{
getMyLastEditRev(.file, .host_name = "mc1.metrumrg.com")
}
\arguments{
\item{.file}{file path from working directory}

\item{.host_name}{Host name where repository is hosted}
}
\description{
Compares the local version of a script with the most recent version the user
committed. This is a helpful tool when reviewing the changes made after QC
comments have been addressed.
}
\examples{
\dontrun{
getMyLastEditRev(.file = "script/data-assembly.Rmd")
}

}
\keyword{internal}
