% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPreviousRevisionDiff.R
\name{getPreviousRevisionDiff}
\alias{getPreviousRevisionDiff}
\title{Visual diff of previous revision of a file to the local version.}
\usage{
getPreviousRevisionDiff(.file, .previous_revision)
}
\arguments{
\item{.file}{file path from working directory}

\item{.previous_revision}{revision to compare to local file}
}
\description{
Compares the local version of a script with a previous revision.
The output will appear in the viewer and only rows where there have been
additions, deletions or modifications in the script will be shown.
}
\examples{
\dontrun{
getPreviousRevisionDiff(.file = "script/data-assembly.Rmd", .previous_revision = 5)
}

}
