% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQcedDiff.R
\name{getQcedDiff}
\alias{getQcedDiff}
\title{Visual diff of last QCed version of a file to the local version.}
\usage{
getQcedDiff(.file)
}
\arguments{
\item{.file}{file path from working directory}
}
\description{
Compares the local version of a script with the most recent QCed version.
The output will appear in the viewer and only rows where there have been
additions, deletions or modifications in the script will be shown.
}
\examples{
\dontrun{
getQcedDiff(.file = "script/data-assembly.Rmd")
}

}
