% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQcedRevision.R
\name{getQcedRevision}
\alias{getQcedRevision}
\title{Determine most recent QCed revision number for a file}
\usage{
getQcedRevision(.file)
}
\arguments{
\item{.file}{file path from working directory}
}
\description{
Returns the most recent revision number of when the specified file was
accepted in the QClog.
}
\examples{
 
\dontrun{
getQcedRevision(.file = "script/data-assembly.Rmd")
}

}
