% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTFLs.R
\name{getTFLs}
\alias{getTFLs}
\title{Generate a list of all tables and figures in a PDF}
\usage{
getTFLs(.file)
}
\arguments{
\item{.file}{File path}
}
\description{
This function parses a PDF containing tables and figures
saved by \code{pmtables} and \code{mrggsave}. It will return a list
of all tables and figures included in the PDF provided.
}
