% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAccept.R
\name{logAccept}
\alias{logAccept}
\title{Approve a File}
\usage{
logAccept(
file=dir(),
directory=getwd(),
origin=logOrigin(file,directory),
reviewer=Sys.info()['user'],
force=FALSE,
...)
}
\arguments{
\item{file}{one or more files to accept}

\item{directory}{the parent of the file}

\item{origin}{the file from which 'file' originates}

\item{reviewer}{the user issuing the opinion, normally oneself}

\item{force}{if false (default) nonexistent files cause an error}

\item{...}{arguments passed to logQueue}
}
\description{
Indicate acceptance of a file by logging an entry with nonzero version.
}
\details{
Can be undone with logRevert(). Calls logAppend(), which gives an error if
the origin is NA (default) and no precedent exists.
}
\author{
Tim Bergsma
}
