% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAssign.R
\name{logAssign}
\alias{logAssign}
\title{Assign Review Tasks}
\usage{
logAssign(
 file = dir(), 
 directory = getwd(), 
 origin = file,
 ...)
}
\arguments{
\item{file}{a character vector of filenames (paths) relative to directory}

\item{directory}{the parent of the file}

\item{origin}{the file from which 'file' originates}

\item{...}{arguments passed to logQueue}
}
\description{
Appends a log with defaults supplied by logQueue().  These are appropriate
for assigments.  Default reviewer is set to "anyone".
Default revision is zero, which is diagnostic of an assignment.
}
\details{
Since files with revision of zero normally do not exist, no acceptance is
implied by an assigned record. This function may be used to specify a reviewer
for certain files.  Default reviewer "anyone" simply declares that an entity
should be reviewed. "Origin" files may also be specified, indicating that a
file originates from (e.g. is created by) some other file.  The function
logPending() identifies files needing (further) review.  A file's origin
is itself, by default. Currently, logAssign() calls logAppend(), which allows
NA for origin (defaults to precedent or to file itself).
}
\examples{
 with_demoRepo({
   logAssign(file = "script/examp-yaml.yaml")
 })

}
\author{
Tim Bergsma
}
