% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logCreate.R
\name{logCreate}
\alias{logCreate}
\title{Create a Review Log}
\usage{
logCreate(directory=getwd())
}
\arguments{
\item{directory}{directory in which to create the log file}
}
\description{
Creates an empty table in the directory specified with name "QClog.csv".
}
\details{
Since files with revision of zero normally do not exist, no acceptance is
implied by an assigned record. This function may be used to specify a reviewer
for certain files.  Default reviewer "anyone" simply declares that an entity
should be reviewed. "Origin" files may also be specified, indicating that a
file originates from (e.g. is created by) some other file.  The function
logPending() identifies files needing (further) review.  A file's origin
is itself, by default. Currently, logAssign() calls logAppend(), which allows
NA for origin (defaults to precedent or to file itself).
}
\author{
Tim Bergsma
}
