% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDelete.R
\name{logDelete}
\alias{logDelete}
\title{Delete file from QClog}
\usage{
logDelete(.filepath)
}
\arguments{
\item{.filepath}{current file path of file to be removed}
}
\description{
This function facilitates deleting a file from the QClog.csv file.
When run, all rows associated with the file will be deleted and
removed from the QC log.
}
