% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRemove.R
\name{logRemove}
\alias{logRemove}
\title{Remove file from QClog}
\usage{
logRemove(.filepath)
}
\arguments{
\item{.filepath}{current file path of file to be removed}
}
\description{
This function facilitates deleting a file from the QClog.csv file.
When run, all rows associated with the file will be deleted and
removed from the QC log.
}
