% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logSummary.R
\name{logSummary}
\alias{logSummary}
\title{Summarize a Review Log}
\usage{
logSummary(directory = getwd())
}
\arguments{
\item{directory}{the directory containing the log, or any directory below it}
}
\value{
A data frame containing one row per file. Columns include \code{file},
\code{origin}, \code{revf}, \code{headf}, \code{revo}, \code{heado},
\code{reviewer}, and \code{time}.
}
\description{
Summarize the review log associated with a particular directory.
}
\details{
Whereas the review log can have any number of records for the same file, the
summary reduces the data to one record per file.  It is an error if any element
of the file column is missing.

The log is read, then sorted on file, revision, time.  The last record
for each file is retained.

The output includes the latest revisions for the file and origin, plus the
reviewer and time information. Times are returned as character strings in
GMT.
}
\examples{
with_demoRepo({
  logSummary()
})

}
\author{
Tim Bergsma
}
