% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myQC.R
\name{myQC}
\alias{myQC}
\title{Outstanding QC check by user}
\usage{
myQC(.user = Sys.info()[["user"]], .max_days = Inf)
}
\arguments{
\item{.user}{A character string indicating the user whose commit history is being examined. Default is the current system user, determined by \code{Sys.info()[["user"]]}.}

\item{.max_days}{A numeric value indicating the maximum number of days since the last edit. Files edited beyond this threshold will be excluded from the results. Default is \code{Inf}, meaning no files are excluded based on the number of days.}
}
\value{
A list of data frames, each representing a set of files with the same QC status. The data frames include the following columns:
\itemize{
\item \code{File}: The name of the file.
\item \verb{Day(s) since last edit}: The number of days since the last modification.
}
}
\description{
This function checks the QC status of files last authored by the given user in a repository.
The information provide includes if a file is awaiting QC or not present in the QC log.
Users can also filter the results by days since their last edit, to only look at files editted
within a specified period of time.
}
\examples{
\dontrun{
# Call myQC for the current user and set a max of 30 days since last edit
result <- myQC(.user = Sys.info()[["user"]], .max_days = 30)

# Print the QC status for files that are awaiting QC
print(result$`Awaiting QC`)
}

}
