% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_revisions.R
\name{next_revisions}
\alias{next_revisions}
\title{Initialise the next review cycle}
\usage{
next_revisions(.data)
}
\arguments{
\item{.data}{A \code{reviewed_df}.}
}
\value{
A \code{claims_df}.
}
\description{
Creates a new review workspace from an accepted review.
}
