% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcComment.R
\name{qcComment}
\alias{qcComment}
\title{Search for QC Comments in Script}
\usage{
qcComment(comment)
}
\arguments{
\item{comment}{Character string containing the QC note}
}
\value{
Invisibly returns the formatted QC comment string
}
\description{
Helper to find lines that use the qcComment() function.
}
\examples{
# In code you would write:
# qcComment("Check outlier threshold")
# review::qcComment("Check outlier threshold") 
}
