% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderOutputs.R
\name{renderOutputs}
\alias{renderOutputs}
\title{Render an Output Mapping Report}
\usage{
renderOutputs(.output_dir, .project_number = NULL)
}
\arguments{
\item{.output_dir}{Character string. Path to the directory where the output PDF
should be saved. This argument is required.}

\item{.project_number}{Character string (optional). Used for the output PDF filename
prefix. If NULL (default), inferred from the SVN project URL.}
}
\description{
This function generates a report showing every file that is output from scripts
in the repo. Only files that have been run with \code{runWithOutputs()} will appear
in the table.
}
\details{
Additionally, there is a check to confirm if the outputs are up to date with
the latest version of the source script.
}
\examples{
\dontrun{
renderOutputs("deliv/report")
}

}
