% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderQCReport.R
\name{renderQCReport}
\alias{renderQCReport}
\title{Render a QC Report Document}
\usage{
renderQCReport(.output_dir, .project_number = NULL)
}
\arguments{
\item{.output_dir}{Character string. Path to the directory where the output PDF
should be saved. This argument is required.}

\item{.project_number}{Character string (optional). Used for the output PDF filename
prefix. If NULL (default), inferred from the SVN project URL.}
}
\value{
Invisible. The function will save a PDF report named "\if{html}{\out{<project>}}-qc-report-(current date).pdf"
to the specified output directory. If the R session is interactive, it will also
open the PDF in the default browser.
}
\description{
This function generates a QC report based on the files within the QC log
and renders it into a PDF report.
}
\details{
The QC status of all files in the QC log will be displayed.
}
\examples{
\dontrun{
renderQCReport("deliv/report")
}

}
