% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoHistory.R
\name{repoHistory}
\alias{repoHistory}
\title{Get Repository History}
\usage{
repoHistory()
}
\description{
Returns a data.frame containing every commit made to the repo.
The information recorded includes the author, revision, date and
commit message.
}
