% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnExport.R
\name{svnExport}
\alias{svnExport}
\title{Export a specific revision of a file from SVN}
\usage{
svnExport(.file, .revision, .output_dir = getwd())
}
\arguments{
\item{.file}{A character string specifying the path to the SVN file to be exported.}

\item{.revision}{A character or numeric value specifying the SVN revision number of the file to be exported.}

\item{.output_dir}{A character string specifying the directory where the exported file should be saved.
Defaults to the current working directory.}
}
\value{
NULL invisibly.
}
\description{
This function exports a specific revision of a file from SVN to a desired output directory.
The exported file will be named with its original base name followed by a hyphen and the revision number.
If the \code{.output_dir} is not specified, the current working directory will be used.
}
\examples{
\dontrun{
  svnExport(.file = "/path/to/svn/file.txt", .revision = "1234", .output_dir = "/path/to/save")
}

}
