% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnInfo.R
\name{svnInfo}
\alias{svnInfo}
\title{Get the SVN info of a file}
\usage{
svnInfo(.file)
}
\arguments{
\item{.file}{File path of the file of interest}
}
\description{
Returns the \verb{svn info} in dataframe format.
}
