% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnMeta.R
\name{svnMeta}
\alias{svnMeta}
\title{Get the SVN history of a file}
\usage{
svnMeta(.file)
}
\arguments{
\item{.file}{Relative file path of the file of interest}
}
\description{
Returns the raw output of \verb{svn log} and \verb{svn info} for the specified
file. \verb{svn log} provides information regarding each revision of the
file and \verb{svn info} provides an overview of the latest update.
}
