% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnProjInfo.R
\name{svnProjInfo}
\alias{svnProjInfo}
\title{Get svn project info}
\usage{
svnProjInfo(.host_name = "mc1.metrumrg.com")
}
\arguments{
\item{.host_name}{Host name where repository is hosted}
}
\description{
Returns info about the svn project
}
\keyword{internal}
