% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnUser.R
\name{svnUser}
\alias{svnUser}
\title{Get local and svn username}
\usage{
svnUser(.host_name = "mc1.metrumrg.com")
}
\arguments{
\item{.host_name}{Host name where repository is hosted}
}
\description{
Returns the user's SVN username, along with the repository root, hostname
and local username.
}
\keyword{internal}
