% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDirQC.R
\name{checkDirQC}
\alias{checkDirQC}
\title{Check if Files in Directory are in QC Log}
\usage{
checkDirQC(.dir_path)
}
\arguments{
\item{.dir_path}{Path to directory containing files to search}
}
\description{
Helper function to ensure all files needing QC are added to the QC log.
Provide a directory and it will be recursively searched for all files that
exist within it. Each file will be checked if it exists in the QC log and if
it is returned by \code{logPending()}.
}
\examples{
\dontrun{
checkDirQC(here::here("script"))
}
}
