% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFigures.R
\name{compareFigures}
\alias{compareFigures}
\title{Generate comparison between sets of figures}
\usage{
compareFigures(
  .path_current,
  .path_previous = NULL,
  .file_types = "pdf",
  .side_by_side = TRUE
)
}
\arguments{
\item{.path_current}{file or directory path to figures of interest}

\item{.path_previous}{file or directory path to compare .path_base to (leave as NULL to compare to repo version)}

\item{.file_types}{file extensions to include in comparison (only pdf and png allowed)}

\item{.side_by_side}{Logical. Should outputs be displayed side by side?}
}
\description{
Generate a html document showing two versions of a figure (pdf or png).
If two paths are provided, the local versions of both will be used.
If only one path is provided, the comparison will show the local
versus last checked in version.
}
