% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLocal.R
\name{compareLocal}
\alias{compareLocal}
\title{Generate comparison between sets of local file(s)}
\usage{
compareLocal(
  .path_base,
  .path_compare,
  .side_by_side = TRUE,
  .file_exts = c("png", "pdf", "tex")
)
}
\arguments{
\item{.path_base}{file or directory path to tables/figures of interest}

\item{.path_compare}{file or directory path to compare .path_base to}

\item{.side_by_side}{Logical. Should outputs be displayed side by side?}

\item{.file_exts}{file extensions to include in comparison}
}
\description{
Generate comparison between sets of local file(s)
}
