% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModified.R
\name{compareModified}
\alias{compareModified}
\title{Generate comparison between local (modified) and checked in file(s)}
\usage{
compareModified(
  .path,
  .path_compare = NULL,
  .side_by_side = TRUE,
  .display_all = FALSE
)
}
\arguments{
\item{.path}{file or directory path to tables/figures of interest (pdf, png, or tex)}

\item{.path_compare}{file or directory path to compare .path to (leave as NULL to compare from svn)}

\item{.side_by_side}{Logical. Should outputs be displayed side by side?}

\item{.display_all}{Logical. Should all outputs start displayed?}
}
\description{
Generate comparison between local (modified) and checked in file(s)
}
