% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffFigure.R
\name{diffFigure}
\alias{diffFigure}
\title{Create a GIF of original vs modified figure}
\usage{
diffFigure(.file1, .file2, .fps = 2)
}
\arguments{
\item{.file1}{file path from working directory}

\item{.file2}{file path from working directory}

\item{.fps}{frames per second, increase for quicker switching between images}
}
\description{
Compares the local version of a figure with the most recent version checked into
SVN. A GIF will be created to help the user compare the original file to the
modified one.
}
\examples{
\dontrun{
diffFigure(
  .file1 = "example-1.pdf",
  .file2 = "example-2.pdf"
  )
}

}
