% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffFiles.R
\name{diffFiles}
\alias{diffFiles}
\title{Visual diff between two different files}
\usage{
diffFiles(
  .file_1,
  .file_2,
  .banner_1 = NULL,
  .banner_2 = NULL,
  .side_by_side = TRUE,
  .ignore_white_space = FALSE
)
}
\arguments{
\item{.file_1}{file path of first file from working directory}

\item{.file_2}{file path of second file from working directory}

\item{.banner_1}{Header for first file in viewer}

\item{.banner_2}{Header for second file in viewer}

\item{.side_by_side}{Logical. Should diffs be displayed side by side?}

\item{.ignore_white_space}{Logical. Should white space be ignored?}
}
\description{
Compares the current version of two scripts against one another.
The output will appear in the viewer and only rows where there have been
additions, deletions or modifications in the script will be shown.

This should be used when a new script has been created based off
an existing script.
}
\examples{
with_demoRepo({
 diffFiles(.file_1 = "script/data-assembly.R", 
           .file_2 = "script/combine-da.R")
})

}
