% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirSummary.R
\name{dirSummary}
\alias{dirSummary}
\title{QC Summary of Files within a Directory}
\usage{
dirSummary(.dir)
}
\arguments{
\item{.dir}{Character string. Path to the directory to perform the QC summary on.}
}
\value{
A list containing:
\itemize{
\item \code{project}: the name of the project repository.
\item \code{data}: a data.frame containing information about the relevant files including:
\itemize{
\item \code{File}: the name of the file.
\item \code{Author}: the last author to modify the file.
\item \verb{Latest edit}: the latest date and time of modification.
\item \verb{Latest rev}: the last revision number.
\item \code{Status}: indicating whether the file is in SVN, in QC log, and if it needs QC.
\item \code{QCer}: the reviewer name if the file is in QC log.
}
\item \code{directory}: the provided directory.
}
}
\description{
This function provides a QC summary of relevant files in a specified directory.
It checks if the file has been checked into SVN, its latest modification date,
author, revision and QC status.
}
\examples{
\dontrun{
# Assuming appropriate setup and the presence of required 
# helper functions like logSummary and svnLog.
# dirSummary("/path/to/directory")
}

}
\seealso{
\code{\link[review]{logSummary}}, \code{\link[review]{svnLog}}
}
