% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSummary.R
\name{fileSummary}
\alias{fileSummary}
\title{Return summary of revision and QC history for files}
\usage{
fileSummary(.file)
}
\arguments{
\item{.file}{Character vectors of file paths}
}
\description{
Returns a summary of the revision and QC history of a file.
Users can provide a single file (or multiple) and will be
provided with file information such as: QC status, if the
file is in the QC log, past authors and past QCers.
}
