% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logSummary.R
\name{logSummary}
\alias{logSummary}
\title{Summarize a Review Log}
\usage{
logSummary(directory = getwd())
}
\arguments{
\item{directory}{the directory containing the log, or any directory below it}
}
\description{
Summarize the review log associated with a particular directory.
}
\details{
Whereas the review log can have any number of records for the same file, the
summary reduces the data to one record per file.  It is an error if any element
of the file column is missing.

The log is read, then sorted on file, revision, time.  The last record
for each file is retained.
}
\author{
Tim Bergsma
}
