% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myQC.R
\name{myQC}
\alias{myQC}
\title{Outstanding QC check by user}
\usage{
myQC(.within_days = Inf)
}
\arguments{
\item{.within_days}{A numeric value indicating how many days back to look for edits.
Default is \code{Inf}, meaning no time limit is applied.}
}
\value{
A list of data frames, each representing a set of files with the same QC status. The data frames include the following columns:
\itemize{
\item \code{File}: The name of the file.
\item \verb{Day(s) since last edit}: The number of days since the last modification.
}
}
\description{
This function checks the QC status of files last authored by the given user in a repository.
The information provided includes if a file is awaiting QC or not present in the QC log.
Users can also filter the results by days since their last edit, to only look at files edited
within a specified period of time.
}
\examples{
\dontrun{
# Call myQC for the current user and look at files edited within the last 14 days
result <- myQC(.within_days = 14)

# Print the QC status for files that are awaiting QC
print(result$`Awaiting QC`)
}

}
\keyword{experimental}
\keyword{internal}
