% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderQCReport.R
\name{renderQCReport}
\alias{renderQCReport}
\title{Render a QC Report Document}
\usage{
renderQCReport(.output_dir = NULL)
}
\arguments{
\item{.output_dir}{Character string (optional). Path to the directory where the output PDF
should be saved. If not provided, the document will not be saved locally.}
}
\value{
Invisible. The function will save a PDF report named "qc-report-(current date).pdf"
to the specified output directory (if available). If the R session is interactive, it will also
open the PDF in the default browser.
}
\description{
This function generates a QC report based on the files within the QC log
and renders it into a PDF report.
}
\details{
The QC status of all files in the QC log will be displayed.
}
\examples{
\dontrun{
renderQCReport()
}

}
