% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewPackage.R
\docType{package}
\name{review-package}
\alias{review}
\alias{review-package}
\title{review: Manage Review Logs}
\description{
Functions for managing logs of reviews of subversioned files (http://subversion.apache.org/).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/metrumresearchgroup/review}
  \item Report bugs at \url{https://github.com/metrumresearchgroup/review/issues}
}

}
\author{
\strong{Maintainer}: Tim Bergsma \email{timb@metrumrg.com}

Authors:
\itemize{
  \item Eric Anderson \email{andersone@metrumrg.com}
  \item Michael McDermott \email{michaelm@metrumrg.com}
}

}
\keyword{internal}
