% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnLog.R
\name{svnLog}
\alias{svnLog}
\title{Get the SVN history of a file}
\usage{
svnLog(.file)
}
\arguments{
\item{.file}{File path of the file of interest}
}
\description{
Returns the \verb{svn log} in dataframe format.
}
